+++ tests/test_util/test_util_image.py
@@ -1,5 +1,5 @@
-from creevey.util.image import find_image_files
 from tests.conftest import TEMP_DATA_DIR
+from wildebeest.util.image import find_image_files
 
 
 def test_find_images(generate_file_tree):  # noqa: F811