+++ wildebeest/__init__.py
@@ -1,0 +1,2 @@
+from wildebeest._version import __version__
+from wildebeest.pipelines import *