+++ wildebeest/load_funcs/__init__.py
@@ -1,0 +1,1 @@
+from wildebeest.load_funcs.load_funcs import *