+++ wildebeest/ops/image/__init__.py
@@ -1,0 +1,2 @@
+from wildebeest.ops.image.stats import *
+from wildebeest.ops.image.transforms import *