+++ wildebeest/path_funcs/__init__.py
@@ -1,0 +1,1 @@
+from wildebeest.path_funcs.path_funcs import *