+++ wildebeest/util/__init__.py
@@ -1,0 +1,1 @@
+from wildebeest.util.util import *