+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 #[3.0.0] - 2020-7-28
+### Changed
+ - Rename library "wildebeest"
+ - Use GitHub Actions for CI/CD.
+
+==== Below is Creevey development ====
+
+#[3.0.0] - 2020-7-28
 ### Added
  - Warning about upcoming name change